refactor tool actors
