+++ agent/nicegui_agent/actors.py
@@ -1,18 +1,16 @@
-import re
 import jinja2
 import logging
-import dataclasses
 import anyio
 from core.base_node import Node
 from core.workspace import Workspace
-from core.actors import BaseData, BaseActor, LLMActor
-from llm.common import AsyncLLM, Message, TextRaw, Tool, ToolUse, ToolUseResult, ThinkingBlock
+from core.actors import BaseData, FileOperationsActor
+from llm.common import AsyncLLM, Message, TextRaw, Tool, ToolUse, ToolUseResult
 from nicegui_agent import playbooks
 
 logger = logging.getLogger(__name__)
 
 
-class NiceguiActor(BaseActor, LLMActor):
+class NiceguiActor(FileOperationsActor):
     root: Node[BaseData] | None = None
 
     def __init__(
@@ -23,15 +21,8 @@ class NiceguiActor(BaseActor, LLMActor):
         max_depth: int = 30,
         system_prompt: str = playbooks.APPLICATION_SYSTEM_PROMPT,
     ):
-        self.llm = llm
-        self.workspace = workspace
-        self.beam_width = beam_width
-        self.max_depth = max_depth
+        super().__init__(llm, workspace, beam_width, max_depth)
         self.system_prompt = system_prompt
-        self.root = None
-        logger.info(
-            f"Initialized {self.__class__.__name__} with beam_width={beam_width}, max_depth={max_depth}"
-        )
 
     async def execute(
         self,
@@ -116,55 +107,9 @@ class NiceguiActor(BaseActor, LLMActor):
         return candidates
 
     @property
-    def tools(self) -> list[Tool]:
+    def additional_tools(self) -> list[Tool]:
+        """NiceGUI-specific tools."""
         return [
-            {
-                "name": "read_file",
-                "description": "Read file content",
-                "input_schema": {
-                    "type": "object",
-                    "properties": {
-                        "path": {"type": "string"},
-                    },
-                    "required": ["path"],
-                },
-            },
-            {
-                "name": "write_file",
-                "description": "Write content to a file",
-                "input_schema": {
-                    "type": "object",
-                    "properties": {
-                        "path": {"type": "string"},
-                        "content": {"type": "string"},
-                    },
-                    "required": ["path", "content"],
-                },
-            },
-            {
-                "name": "edit_file",
-                "description": "Edit a file by searching and replacing text",
-                "input_schema": {
-                    "type": "object",
-                    "properties": {
-                        "path": {"type": "string"},
-                        "search": {"type": "string"},
-                        "replace": {"type": "string"},
-                    },
-                    "required": ["path", "search", "replace"],
-                },
-            },
-            {
-                "name": "delete_file",
-                "description": "Delete a file",
-                "input_schema": {
-                    "type": "object",
-                    "properties": {
-                        "path": {"type": "string"},
-                    },
-                    "required": ["path"],
-                },
-            },
             {
                 "name": "uv_add",
                 "description": "Install additional packages",
@@ -176,215 +121,37 @@ class NiceguiActor(BaseActor, LLMActor):
                     "required": ["packages"],
                 },
             },
-            {
-                "name": "complete",
-                "description": "Mark the task as complete. This will run tests and type checks to ensure the changes are correct.",
-                "input_schema": {
-                    "type": "object",
-                    "properties": {},
-                },
-            },
         ]
 
-    async def run_tools(
-        self, node: Node[BaseData], user_prompt: str
-    ) -> tuple[list[ToolUseResult], bool]:
-        logger.info(f"Running tools for node {node._id}")
-        result, is_completed = [], False
-        for block in node.data.head().content:
-            if not isinstance(block, ToolUse):
-                match block:
-                    case TextRaw(text=text):
-                        logger.info(f"LLM output: {text}")
-                    case _:
-                        pass
-                continue
-            try:
-
-                def _short_dict_repr(d: dict) -> str:
-                    return ", ".join(
-                        f"{k}: {v if len(v) < 100 else v[:50] + '...'}"
-                        for k, v in d.items()
-                        if isinstance(v, str)
-                    )
-
-                logger.info(
-                    f"Running tool {block.name} with input {_short_dict_repr(block.input) if isinstance(block.input, dict) else str(block.input)}"
+    async def handle_custom_tool(
+        self, tool_name: str, tool_input: dict, node: Node[BaseData]
+    ) -> ToolUseResult:
+        """Handle NiceGUI-specific custom tools."""
+        match tool_name:
+            case "uv_add":
+                packages = tool_input["packages"]  # pyright: ignore[reportIndexIssue]
+                exec_res = await node.data.workspace.exec_mut(
+                    ["uv", "add", " ".join(packages)]
                 )
-
-                match block.name:
-                    case "read_file":
-                        tool_content = await node.data.workspace.read_file(
-                            block.input["path"]  # pyright: ignore[reportIndexIssue]
-                        )
-                        result.append(ToolUseResult.from_tool_use(block, tool_content))
-                    case "write_file":
-                        path = block.input["path"]  # pyright: ignore[reportIndexIssue]
-                        content = block.input["content"]  # pyright: ignore[reportIndexIssue]
-                        try:
-                            node.data.workspace.write_file(path, content)
-                            node.data.files.update({path: content})
-                            result.append(ToolUseResult.from_tool_use(block, "success"))
-                            logger.debug(f"Written file: {path}")
-                        except FileNotFoundError as e:
-                            error_msg = (
-                                f"Directory not found for file '{path}': {str(e)}"
+                if exec_res.exit_code != 0:
+                    return ToolUseResult.from_tool_use(
+                        ToolUse(id="", name=tool_name, input=tool_input),
+                        f"Failed to add packages: {exec_res.stderr}",
+                        is_error=True,
+                    )
+                else:
+                    node.data.files.update(
+                        {
+                            "pyproject.toml": await node.data.workspace.read_file(
+                                "pyproject.toml"
                             )
-                            logger.info(
-                                f"File not found error writing file {path}: {str(e)}"
-                            )
-                            result.append(
-                                ToolUseResult.from_tool_use(
-                                    block, error_msg, is_error=True
-                                )
-                            )
-                        except PermissionError as e:
-                            error_msg = (
-                                f"Permission denied writing file '{path}': {str(e)}"
-                            )
-                            logger.info(
-                                f"Permission error writing file {path}: {str(e)}"
-                            )
-                            result.append(
-                                ToolUseResult.from_tool_use(
-                                    block, error_msg, is_error=True
-                                )
-                            )
-                        except ValueError as e:
-                            error_msg = str(e)
-                            logger.info(f"Value error writing file {path}: {error_msg}")
-                            result.append(
-                                ToolUseResult.from_tool_use(
-                                    block, error_msg, is_error=True
-                                )
-                            )
-                    case "edit_file":
-                        path = block.input["path"]  # pyright: ignore[reportIndexIssue]
-                        search = block.input["search"]  # pyright: ignore[reportIndexIssue]
-                        replace = block.input["replace"]  # pyright: ignore[reportIndexIssue]
-
-                        try:
-                            original = await node.data.workspace.read_file(path)
-                            match original.count(search):
-                                case 0:
-                                    raise ValueError(
-                                        f"Search text not found in file '{path}'. Search:\n{search}"
-                                    )
-                                case 1:
-                                    new_content = original.replace(search, replace)
-                                    node.data.workspace.write_file(path, new_content)
-                                    node.data.files.update({path: new_content})
-                                    result.append(
-                                        ToolUseResult.from_tool_use(block, "success")
-                                    )
-                                    logger.debug(f"Applied edit to file: {path}")
-                                case num_hits:
-                                    raise ValueError(
-                                        f"Search text found {num_hits} times in file '{path}' (expected exactly 1). Search:\n{search}"
-                                    )
-                        except FileNotFoundError as e:
-                            error_msg = f"File '{path}' not found for editing: {str(e)}"
-                            logger.info(
-                                f"File not found error editing file {path}: {str(e)}"
-                            )
-                            result.append(
-                                ToolUseResult.from_tool_use(
-                                    block, error_msg, is_error=True
-                                )
-                            )
-                        except PermissionError as e:
-                            error_msg = (
-                                f"Permission denied editing file '{path}': {str(e)}"
-                            )
-                            logger.info(
-                                f"Permission error editing file {path}: {str(e)}"
-                            )
-                            result.append(
-                                ToolUseResult.from_tool_use(
-                                    block, error_msg, is_error=True
-                                )
-                            )
-                        except ValueError as e:
-                            error_msg = str(e)
-                            logger.info(f"Value error editing file {path}: {error_msg}")
-                            result.append(
-                                ToolUseResult.from_tool_use(
-                                    block, error_msg, is_error=True
-                                )
-                            )
-                    case "delete_file":
-                        node.data.workspace.rm(block.input["path"])  # pyright: ignore[reportIndexIssue]
-                        node.data.files.update({block.input["path"]: None})  # pyright: ignore[reportIndexIssue]
-                        result.append(ToolUseResult.from_tool_use(block, "success"))
-                    case "uv_add":
-                        packages = block.input["packages"]  # pyright: ignore[reportIndexIssue]
-                        exec_res = await node.data.workspace.exec_mut(
-                            ["uv", "add", " ".join(packages)]
-                        )
-                        if exec_res.exit_code != 0:
-                            result.append(
-                                ToolUseResult.from_tool_use(
-                                    block,
-                                    f"Failed to add packages: {exec_res.stderr}",
-                                    is_error=True,
-                                )
-                            )
-                        else:
-                            node.data.files.update(
-                                {
-                                    "pyproject.toml": await node.data.workspace.read_file(
-                                        "pyproject.toml"
-                                    )
-                                }
-                            )
-                            result.append(ToolUseResult.from_tool_use(block, "success"))
-                    case "complete":
-                        if not self.has_modifications(node):
-                            raise ValueError(
-                                "Can not complete without writing any changes."
-                            )
-                        check_err = await self.run_checks(node, user_prompt)
-
-                        if check_err:
-                            logger.info(f"Failed to complete: {check_err}")
-                        result.append(
-                            ToolUseResult.from_tool_use(block, check_err or "success")
-                        )
-                        node.data.should_branch = True
-                        is_completed = check_err is None
-
-                    case unknown:
-                        raise ValueError(f"Unknown tool: {unknown}")
-            except FileNotFoundError as e:
-                logger.info(f"File not found: {e}")
-                result.append(ToolUseResult.from_tool_use(block, str(e), is_error=True))
-            except PermissionError as e:
-                logger.info(f"Permission error: {e}")
-                result.append(ToolUseResult.from_tool_use(block, str(e), is_error=True))
-            except ValueError as e:
-                logger.info(f"Value error: {e}")
-                result.append(ToolUseResult.from_tool_use(block, str(e), is_error=True))
-            except Exception as e:
-                logger.error(f"Unknown error: {e}")
-                result.append(ToolUseResult.from_tool_use(block, str(e), is_error=True))
-        return result, is_completed
-
-    async def eval_node(self, node: Node[BaseData], user_prompt: str) -> bool:
-        tool_calls, is_completed = await self.run_tools(node, user_prompt)
-        if tool_calls:
-            node.data.messages.append(Message(role="user", content=tool_calls))
-        else:
-            content = [TextRaw(text="Continue or mark completed.")]
-            node.data.messages.append(Message(role="user", content=content))
-        return is_completed
-
-    def has_modifications(self, node: Node[BaseData]) -> bool:
-        cur_node = node
-        while cur_node is not None:
-            if cur_node.data.files:
-                return True
-            cur_node = cur_node.parent
-        return False
+                        }
+                    )
+                    return ToolUseResult.from_tool_use(
+                        ToolUse(id="", name=tool_name, input=tool_input), "success"
+                    )
+            case _:
+                return await super().handle_custom_tool(tool_name, tool_input, node)
 
     async def run_type_checks(self, node: Node[BaseData]) -> str | None:
         type_check_result = await node.data.workspace.exec(
@@ -492,15 +259,3 @@ class NiceguiActor(BaseActor, LLMActor):
             ]:
                 repo_files.add(file_path)
         return sorted(list(repo_files))
-
-    async def dump(self) -> object:
-        if self.root is None:
-            return []
-        return await self.dump_node(self.root)
-
-    async def load(self, data: object):
-        if not data:
-            return
-        if not isinstance(data, list):
-            raise ValueError(f"Expected list got {type(data)}")
-        self.root = await self.load_node(data)
